requirements.txt
torch~=2.0.0
argparse~=1.4.0
transformers~=4.29.1
utils~=1.0.1
tqdm~=4.64.1
torchvision~=0.15.1
chardet~=5.1.0
Pillow~=8.4.0
numpy~=1.22.4
sklearn~=0.0
scikit-learn~=1.0.1